+++ src/main/java/cart/controller/dto/ProductUpdateRequest.java
@@ -7,7 +7,7 @@ import javax.validation.constraints.NotNull;
 
 public class ProductUpdateRequest {
     @NotNull(message = "아이디가 비어있습니다.")
-    private final Integer id;
+    private final Long id;
     @NotBlank(message = "이름이 비어있습니다.")
     private String name;
     @NotBlank(message = "이미지가 비어있습니다.")
@@ -17,16 +17,16 @@ public class ProductUpdateRequest {
             max = 100000000,
             message = "상품의 가격은 0~100000000 사이의 값이여야 합니다."
     )
-    private Integer price;
+    private Long price;
 
-    public ProductUpdateRequest(final Integer id, final String name, final String image, final Integer price) {
+    public ProductUpdateRequest(final Long id, final String name, final String image, final Long price) {
         this.id = id;
         this.name = name;
         this.image = image;
         this.price = price;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -38,7 +38,7 @@ public class ProductUpdateRequest {
         return image;
     }
 
-    public Integer getPrice() {
+    public Long getPrice() {
         return price;
     }
 }
